requirements.txt
pandas
mlflow==2.11.3
notebook
numpy
scikit-learn==1.5.0
matplotlib
python-box==7.2.0
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
tensorflow
keras
imbalanced-learn==0.12.3
lightgbm
scipy
seaborn
umap-learn
xgboost
ctgan
torch
apache-airflow==2.9.2
openpyxl==3.1.5

-e .
